backend/agricultural-labor-management/requirements.txt
flask  
flask-pymongo  
flask-wtf  
flask-session  